+++ packages/html-reporter/src/reportView.tsx
@@ -23,6 +23,7 @@ import { HeaderView } from './headerView';
 import { Route } from './links';
 import type { LoadedReport } from './loadedReport';
 import './reportView.css';
+import type { GitCommitInfo } from './metadataView';
 import { MetadataView } from './metadataView';
 import { TestCaseView } from './testCaseView';
 import { TestFilesView } from './testFilesView';
@@ -47,7 +48,7 @@ export const ReportView: React.FC<{
   return <div className='htmlreport vbox px-4 pb-4'>
     <main>
       {report?.json() && <HeaderView stats={report.json().stats} filterText={filterText} setFilterText={setFilterText}></HeaderView>}
-      {report?.json().metadata && <HideIfErrors><MetadataView {...report?.json().metadata!} /></HideIfErrors>}
+      {report?.json().metadata && <HideIfErrors><MetadataView {...report?.json().metadata['git-commit-info'] as GitCommitInfo} /></HideIfErrors>}
       <Route params=''>
         <TestFilesView report={report?.json()} filter={filter} expandedFiles={expandedFiles} setExpandedFiles={setExpandedFiles}></TestFilesView>
       </Route>
